Finished the reports page, reorganised scripts and include items
