+++ pages/factory.php
@@ -9,12 +9,7 @@
 </head>
 <body>
     <?php
-        require_once '../include/functions.php';
-        require_once '../include/database.php';
-        require_once '../include/check-authorisation.php';
-        checkMachineIdIsSet($conn);
-        require_once '../include/menu.php';
-        setBannerColour($conn);
+        require_once '../include/page-defaults.php';
         mysqli_close($conn);
     ?>
     <div id=body-container>
