+++ pages/login.php
@@ -12,10 +12,11 @@
 </head>
 <body>
     <?php
-        require_once '../include/functions.php';
+        require_once '../include/utilities.php';
         require_once '../include/database.php';
         redirectToDashboardIfLoggedIn();
         checkMachineIdIsSet($conn);
+        require_once '../scripts/login.php';
     ?>
     <div id=header-container>
         <p id=header-message>Smart Manafacturing Dashboard</p>
