+++ pages/reports.php
@@ -6,88 +6,78 @@
     <meta name="author" content="Group 18" />
     <link rel="stylesheet" href="../styles/style.css">
     <script src="../scripts/banner.js"></script>
-    <script src="../scripts/reports-buttons.js" defer></script>
+    <script src="../scripts/reports.js" defer></script>
 </head>
 <body>
     <?php
-        require_once '../include/functions.php';
-        require_once '../include/database.php';
-        require_once '../include/check-authorisation.php';
-        checkMachineIdIsSet($conn);
-        require_once '../include/menu.php';
-        setBannerColour($conn);
+        require_once '../include/page-defaults.php';
+        require_once '../scripts/reports.php';
     ?>
-    <div id=body-container>
+    <div id="body-container">
         <h1>Reports</h1>
-        <div id=reports-button-container>
-            <div id=reports-button-select-start-wrapper class=reports-button>
+        <form id="reports-button-container" action="reports.php" method="GET">
+            <input type="hidden" name="machineID" value="<?php echo $_GET['machineID'] ?>"/>  <!-- This is a bit ugly, it's there because submitting a GET form erases the existing query parameters.  -->
+            <div id="reports-button-input-start-wrapper" class=reports-button>
                 <p>Start date: </p>
-                <input id=reports-button-select-start name="start" type="datetime-local"/>
+                <input id="reports-button-input-start" class="reports-input" name="start" type="datetime-local" required/>
             </div>
-            <div id=reports-button-select-end-wrapper class=reports-button>
+            <div id="reports-button-input-end-wrapper" class=reports-button>
                 <p>End date: </p>
-                <input id=reports-button-select-end name="end" type="datetime-local"/>
+                <input id="reports-button-input-end" class="reports-input" name="end" type="datetime-local" required/>
             </div>
-            <input id=reports-button-submit name="submit" type="submit" value="Submit"/>
+            <select id="reports-button-select-machine" name="machine" required>
+                <?php 
+                    appendMachinesToSelect($conn);
+                ?>
+            </select>
+            <input id="reports-button-submit" type="submit" value="✓"/>
+        </form>
+        <div id="reports-table-wrapper">
+            <table>
+                <tr id="reports-labels">
+                    <th>
+                        Timestamp
+                    </th>
+                    <th>
+                        Machine Name
+                    </th>
+                    <th>
+                        Status
+                    </th>
+                    <th>
+                        Log
+                    </th>
+                    <th>
+                        Error
+                    </th>
+                    <th>
+                        Production Count
+                    </th>
+                    <th>
+                        Humidity
+                    </th>
+                    <th>
+                        Power Consumption
+                    </th>
+                    <th>
+                        Pressure
+                    </th>
+                    <th>
+                        Speed
+                    </th>
+                    <th>
+                        Temperature
+                    </th>
+                    <th>
+                        Vibration
+                    </th>
+                </tr>
+                <?php 
+                getLogs($conn);
+                mysqli_close($conn);
+                ?>
+            </table>
         </div>
-        <table id=reports-table>
-            <tr id=reports-labels>
-                <th>
-                    Timestamp
-                </th>
-                <th>
-                    Machine Name
-                </th>
-                <th>
-                    Status
-                </th>
-                <th>
-                    Log
-                </th>
-                <th>
-                    Error
-                </th>
-                <th>
-                    Production Count
-                </th>
-                <th>
-                    Humidity
-                </th>
-                <th>
-                    Power Consumption
-                </th>
-                <th>
-                    Pressure
-                </th>
-                <th>
-                    Speed
-                </th>
-                <th>
-                    Temperature
-                </th>
-                <th>
-                    Vibration
-                </th>
-            </tr>
-            <?php 
-            $machine_assoc = getMachineNames($conn);
-            $timestamp_start = '2024-06-30 00:00';
-            $timestamp_end = '2024-06-30 23:30';
-            $sql = "SELECT * FROM Log WHERE timestamp >= \"$timestamp_start\" AND timestamp <= \"$timestamp_end\";";
-            $result = mysqli_query($conn, $sql);
-            if ($result && mysqli_num_rows($result)) {
-                while ($assoc = mysqli_fetch_assoc($result)) {
-                    $assoc['name'] = $machine_assoc[$assoc['machineID']];
-                    appendLogToTable($assoc);
-                }
-            }
-            else {
-                echo 'Unable to query logs. ';
-            }
-            mysqli_free_result($result);
-            mysqli_close($conn);
-            ?>
-        </table>
     </div>
 </body>
 </html>
