+++ pages/unauthorised.php
@@ -9,8 +9,9 @@
 </head>
 <body>
     <?php
-        require_once '../include/functions.php';
+        require_once '../include/utilities.php';
         require_once '../include/database.php';
+        require_once '../scripts/unauthorised.php';
     ?>
     <div id=header-container>
         <p id=header-message>Smart Manafacturing Dashboard</p>
