+++ pages/users.php
@@ -9,12 +9,8 @@
 </head>
 <body>
     <?php
-        require_once '../include/functions.php';
-        require_once '../include/database.php';
-        require_once '../include/check-authorisation.php';
-        checkMachineIdIsSet($conn);
-        require_once '../include/menu.php';
-        setBannerColour($conn);
+        require_once '../include/page-defaults.php';
+        require_once '../scripts/users.php';
     ?>
     <div id="body-container">
         <h1>Users</h1>
