+++ scripts/unauthorised.php
@@ -1,0 +1,10 @@
+<?php 
+function setUnauthorisedButton() {
+    if (isset($_SESSION['home'])) {
+        echo "<a href=\"{$_SESSION['home']}?machineID={$_GET['machineID']}\">Home</a>";
+    }
+    else {
+        echo '<a href="login.php">Login</a>';
+        session_destroy();
+    }
+}
