+++ scripts/users.php
@@ -1,0 +1,20 @@
+<?php 
+function appendUserToList($assoc) {
+    echo '<li>';
+        echo '<div>'.$assoc['firstName'].' '.$assoc['lastName'].'</div>';
+        echo '<table>';
+            echo '<tr>';
+                echo '<td>Position: '.$assoc['position'].'</td>';
+                echo '<td>User ID: '.$assoc['personID'].'</td>';
+            echo '</tr>';
+            echo '<tr>';
+                echo '<td>Phone: '.$assoc['phoneNumber'].'</td>';
+                echo '<td>Email: '.$assoc['email'].'</td>';
+            echo '</tr>';
+            echo '<tr>';
+                echo '<td>DOB: '.$assoc['DOB'].'</td>';
+                echo '<td>Start Date: '.$assoc['employmentDate'].'</td>';
+            echo '</tr>';
+        echo '</table>';
+    echo '</li>';
+}
